ECS: Move rendering logic to system
